Make parts of the test infrastructure reusable
In that manner, the relevant code has not to be duplicated into
"mqttwarn-contrib".
