+++ tests/test_pushover.py
@@ -7,7 +7,7 @@ import responses
 from requests_toolbelt import MultipartDecoder
 
 from mqttwarn.util import load_module_from_file
-from tests.util import ProcessorItem as Item
+from mqttwarn.model import ProcessorItem as Item
 
 
 def add_successful_mock_response():
