+++ tests/util.py
@@ -1,8 +1,6 @@
 # -*- coding: utf-8 -*-
 # (c) 2018-2021 The mqttwarn developers
 import time
-from dataclasses import dataclass
-from typing import Dict, Union, List
 
 from paho.mqtt.client import MQTTMessage
 
@@ -41,20 +39,3 @@ def send_message(topic=None, payload=None):
 
     # Give the machinery some time to process the message
     time.sleep(0.10)
-
-
-@dataclass
-class ProcessorItem:
-    """
-    A surrogate processor item for feeding into service handlers.
-    """
-
-    service: str = None
-    target: str = None
-    config: Dict = None
-    addrs: List[str] = None
-    priority: int = None
-    topic: str = None
-    title: str = None
-    message: Union[str, bytes] = None
-    data: Dict = None
